Add collision detection and solution
Use OBB to describe each collider (which seems to be too simple and is
not the best idea to describe a sphere).
The solution of a collision is simply to revoke the translation or
rotation of the tank.
The barrel of a tank and the terrain won't trigger the collision
(otherwise the tank cannot move!)